+++ cairos/gui/__init__.py
@@ -1,0 +1,2 @@
+"""Optional local web GUI for CAIROS."""
+